+++ nRFMeshProvision/Classes/Mesh Messages/Lighting/LightLCOccupancyModeSetUnacknowledged.swift
@@ -1,5 +1,5 @@
 /*
-* Copyright (c) 2019, Nordic Semiconductor
+* Copyright (c) 2021, Nordic Semiconductor
 * All rights reserved.
 *
 * Redistribution and use in source and binary forms, with or without modification,
@@ -29,32 +29,30 @@
 */
 
 import Foundation
-import nRFMeshProvision
 
-extension UInt16 {
+public struct LightLCOccupancyModeSetUnacknowledged: GenericMessage {
+    public static let opCode: UInt32 = 0x8297
     
-    // Bluetooth SIG Models
-    static let configurationServerModelId: UInt16 = 0x0000
-    static let configurationClientModelId: UInt16 = 0x0001
+    /// Whether the controller may transition from a standby state when occupancy
+    /// is reported.
+    let occupancyMode: Bool
     
-    static let genericOnOffServerModelId: UInt16 = 0x1000
-    static let genericOnOffClientModelId: UInt16 = 0x1001
-    static let genericLevelServerModelId: UInt16 = 0x1002
-    static let genericLevelClientModelId: UInt16 = 0x1003
+    public var parameters: Data? {
+        return Data() + occupancyMode
+    }
     
-    static let genericDefaultTransitionTimeServerModelId: UInt16 = 0x1004
-    static let genericDefaultTransitionTimeClientModelId: UInt16 = 0x1005
+    /// Creates the Light LC Occupancy Mode Set Unacknowledged message.
+    ///
+    /// - parameter mode: The present value of the Light LC Occupancy Mode state.
+    public init(_ mode: Bool) {
+        self.occupancyMode = mode
+    }
     
-    static let sceneServerModelId: UInt16 = 0x1203
-    static let sceneSetupServerModelId: UInt16 = 0x1204
-    static let sceneClientModelId: UInt16 = 0x1205
-    
-    static let sensorServerModelId: UInt16 = 0x1100
-    static let sensorServerSetupModelId: UInt16 = 0x1101
-    static let sensorClientModelId: UInt16 = 0x1102
-    
-    // Supported vendor models
-    static let simpleOnOffModelId: UInt16 = 0x0001
-    static let nordicSemiconductorCompanyId: UInt16 = 0x0059
+    public init?(parameters: Data) {
+        guard parameters.count == 1 else {
+            return nil
+        }
+        self.occupancyMode = parameters[0] == 0x01
+    }
     
 }