+++ nRFMeshProvision/Classes/Mesh Messages/Sensors/SensorSettingSetUnacknowledged.swift
@@ -1,5 +1,5 @@
 /*
-* Copyright (c) 2019, Nordic Semiconductor
+* Copyright (c) 2021, Nordic Semiconductor
 * All rights reserved.
 *
 * Redistribution and use in source and binary forms, with or without modification,
@@ -30,22 +30,49 @@
 
 import Foundation
 
-public struct LightHSLTargetGet: AcknowledgedGenericMessage {
-    public static let opCode: UInt32 = 0x8279
-    public static let responseType: StaticMeshMessage.Type = LightHSLTargetStatus.self
+public struct SensorSettingSetUnacknowledged: SensorPropertyMessage {
+    public static let opCode: UInt32 = 0x5A
+    
+    public let property: DeviceProperty
+    /// Setting Property identifying a setting within a sensor.
+    public let settingProperty: DeviceProperty
+    /// The value of the setting.
+    public let settingValue: DevicePropertyCharacteristic
     
     public var parameters: Data? {
-        return nil
+        return Data() + property.id + settingProperty.id + settingValue.data
     }
     
-    public init() {
-        // Empty
+    /// Creates the Sensor Setting Set Unacknowledged message.
+    ///
+    /// - parameters:
+    ///   - setting:  Setting Property identifying a setting within a sensor.
+    ///   - property: Property identifying a sensor.
+    ///   - value:    The value of the setting.
+    public init(_ setting: DeviceProperty, of property: DeviceProperty,
+                to value: DevicePropertyCharacteristic) {
+        self.property = property
+        self.settingProperty = setting
+        self.settingValue = value
     }
     
     public init?(parameters: Data) {
-        guard parameters.isEmpty else {
+        guard parameters.count >= 5 else {
             return nil
         }
+        let propertyId: UInt16 = parameters.read(fromOffset: 0)
+        self.property = DeviceProperty(propertyId)
+        
+        let settingPropertyId: UInt16 = parameters.read(fromOffset: 2)
+        self.settingProperty = DeviceProperty(settingPropertyId)
+        
+        // For known properties, make sure we have enough of data to parse value from.
+        if let expectedValueLength = settingProperty.valueLength {
+            guard expectedValueLength == parameters.count - 4 else {
+                return nil
+            }
+        }
+        self.settingValue = settingProperty.read(from: parameters, at: 4, length: parameters.count - 4)
     }
     
 }